requirements.txt
fastapi==0.104.1
fastapi-sessions==0.3.2
SQLAlchemy==2.0.23
uvicorn[standard]==0.24.0.post1
pydantic==2.5.2
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
httpx==0.25.2
asyncpg==0.29.0
greenlet==3.0.1